draft: first version of NeoCalc for Android
This is a experiment trying to use the work previously done on the calculator engine
with a calculator GUI built in Kotlin using Compose. This is supposed to follow the
Material Desgin UI 3 guidelines, however it is not in a good state and should be refined
a lot. We aim the GUI to help provide a similar level experience of the GTK one but not feeling
foreign to the Android ecosystem in the near future.

Signed-off-by: Nilton Perim Neto <niltonperimneto@gmail.com>